refactor(sdk): align task module names with class names
Before this commit, the V1 module `sdk/task.py` held `class TaskConfig`
while the V2 module `sdk/task_config.py` held `class Task` — module and
class names were swapped between the two, making "task vs task_config"
ambiguous on every import.

Rename to align:
- `sdk/task.py` (V1) -> `sdk/task_v1.py` (legacy, marked for deletion
  once `cli/api.py:cli_estimate` migrates off it in a later PR)
- `sdk/task_config.py` (V2) -> `sdk/task.py` (matches `class Task`)

Test directories follow:
- `tests/unit/sdk/task/` (V1 tests) -> `tests/unit/sdk/task_v1/`
- `tests/unit/sdk/task_config/` (V2 tests) -> `tests/unit/sdk/task/`

All import sites updated mechanically:
- 11 files importing V1 -> `from aiconfigurator.sdk.task_v1 import ...`
- 3 files importing V2 -> `from aiconfigurator.sdk.task import ...`

Zero behavior change.  Integration parity (5 cases) and 1057 unit tests
pass.  Phase 2 (cli_integration branch) will pick this up on rebase and
its imports will simplify from `sdk.task_config` to `sdk.task`.
