+++ src/aiconfigurator/generator/module_bridge.py
@@ -8,7 +8,7 @@ from typing import Any
 import pandas as pd
 
 from aiconfigurator.sdk.perf_database import get_database
-from aiconfigurator.sdk.task import TaskConfig
+from aiconfigurator.sdk.task_v1 import TaskConfig
 
 from .aggregators import collect_generator_params
 from .rendering import apply_defaults
