+++ tests/unit/sdk/task_v1/test_task.py
@@ -30,9 +30,9 @@ _SRC = _find_repo_root(pathlib.Path(__file__)) / "src"
 if str(_SRC) not in sys.path:
     sys.path.insert(0, str(_SRC))
 
-import aiconfigurator.sdk.task as task_module
+import aiconfigurator.sdk.task_v1 as task_module
 from aiconfigurator.sdk.errors import NoFeasibleConfigError
-from aiconfigurator.sdk.task import TaskConfig, TaskRunner
+from aiconfigurator.sdk.task_v1 import TaskConfig, TaskRunner
 
 
 @pytest.fixture(autouse=True)
@@ -924,7 +924,7 @@ def test_vllm_disagg_does_not_require_same_tp(monkeypatch):
 # PD-split independent WideEP / EPLB configuration
 # ---------------------------------------------------------------------------
 
-from aiconfigurator.sdk.task import build_disagg_parallel_lists
+from aiconfigurator.sdk.task_v1 import build_disagg_parallel_lists
 
 
 class TestBuildDisaggParallelListsMixedWideep:
